requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
